Change bilinear and cubic interpolation near borders
At the borders "near" zero (top and left) it is doing an extrapolation
instead of downgrading to a simpler interpolation (down to nearest neighbour)
This is not consistent with the right and bottom borders.

This impacts test_transformer_14
The inverse RPC algorithm starts always in pixel 0,0 in this test
due to MIN/MAX_LAT/LON = 0, so it is always affected by this change.

test_transformer_14 is testing that the inverse transformer is converging
when it has troubles converging.
It must include "Oscillation detected." and "Converged!" traces in the debug output.